+++ utils/encryption/encryptor_test.go
@@ -1,4 +1,4 @@
-package rsa_encryptor
+package encryption
 
 import (
 	_ "embed"
